# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bank Statement Analyzer API",
        "description": "API for analyzing bank statements using AI-powered transaction extraction and classification. Supports PDF bank statement uploads, transaction analysis, MCA (Merchant Cash Advance) detection, and revenue classification.",
        "contact": {
            "name": "CRMFinity Support",
            "email": "support@crmfinity.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server v1"
        }
    ],
    "paths": {
...
        "/bank-statement/sessions": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Get all analysis sessions",
                "description": "Retrieve a paginated list of all bank statement analysis sessions.",
                "operationId": "877b1c69afe7be52769411b593bd07e7",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sessions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SessionSummary"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/bank-statement/sessions/{sessionId}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Get session details",
                "description": "Retrieve detailed information about a specific analysis session.",
                "operationId": "e649abd7923cff85295fbad4378ca1f4",
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "description": "Session UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SessionDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Delete a session",
                "description": "Delete an analysis session and all its transactions.",
                "operationId": "b5376f3db6010ddfbea24e2b5aebe75c",
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "description": "Session UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Session deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response",
                "properties": {
                    "success": {
                        "description": "OpenAPI Schema definitions for Bank Statement Analyzer API",
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Error message here"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "title": "Pagination",
                "description": "Pagination metadata",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
...
            "SessionSummary": {
                "title": "Session Summary",
                "description": "Summary of an analysis session",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "session_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "filename": {
                        "type": "string",
                        "example": "bank_statement_jan_2024.pdf"
                    },
                    "pages": {
                        "type": "integer",
                        "example": 5
                    },
                    "total_transactions": {
                        "type": "integer",
                        "example": 150
                    },
                    "total_credits": {
                        "type": "number",
                        "format": "float",
                        "example": 75000
                    },
                    "total_debits": {
                        "type": "number",
                        "format": "float",
                        "example": 60000
                    },
                    "net_flow": {
                        "type": "number",
                        "format": "float",
                        "example": 15000
                    },
                    "analysis_type": {
                        "type": "string",
                        "example": "openai"
                    },
                    "model_used": {
                        "type": "string",
                        "example": "gpt-4o"
                    },
                    "api_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 0.15
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "type": "object"
            },
            "SessionDetail": {
                "title": "Session Detail",
                "description": "Detailed information about an analysis session",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SessionSummary"
                    },
                    {
                        "properties": {
                            "user_id": {
                                "type": "integer",
                                "example": 1,
                                "nullable": true
                            },
                            "high_confidence_count": {
                                "type": "integer",
                                "example": 140
                            },
                            "medium_confidence_count": {
                                "type": "integer",
                                "example": 8
                            },
                            "low_confidence_count": {
                                "type": "integer",
                                "example": 2
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "oauth2",
                "description": "Login with email and password to get access token",
                "flows": {
                    "password": {
                        "tokenUrl": "/api/v1/auth/login",
                        "scopes": {}
                    }
                }
            }
        }
    },
...
}